separated WorkerMiddleware into Worker and MessageRouterMiddleware
