+++ src/Middleware/WorkerMiddleware.php
@@ -28,8 +28,8 @@ use ZfrEbWorker\Exception\RuntimeException;
 
 /**
  * Worker middleware
- * What this thing does is extracting the message from the request, and dispatching a pipeline of the mapped
- * middlewares. You can find a complete reference of what Elastic Beanstalk set here:
+ * What this thing does is validating worker request and extracts message attributes
+ * You can find a complete reference of what Elastic Beanstalk set here:
  * http://docs.aws.amazon.com/elasticbeanstalk/latest/dg/using-features-managing-env-tiers.html
  *
  * @author Michaël Gallego
@@ -42,31 +42,6 @@ class WorkerMiddleware implements MiddlewareInterface
     const MESSAGE_PAYLOAD_ATTRIBUTE      = 'worker.message_payload';
     const MATCHED_QUEUE_ATTRIBUTE        = 'worker.matched_queue';
     const LOCALHOST_ADDRESSES            = ['127.0.0.1', '::1'];
-
-    /**
-     * @var ContainerInterface
-     */
-    private $container;
-
-    /**
-     * Map message names to a middleware name. For instance:
-     * [
-     *     'image.saved' => ProcessImageMiddleware::class
-     * ]
-     *
-     * @var array
-     */
-    private $messagesMapping;
-
-    /**
-     * @param array              $messagesMapping
-     * @param ContainerInterface $container
-     */
-    public function __construct(array $messagesMapping, ContainerInterface $container)
-    {
-        $this->messagesMapping = $messagesMapping;
-        $this->container       = $container;
-    }
 
     /**
      * @param ServerRequestInterface $request
@@ -103,47 +78,7 @@ class WorkerMiddleware implements MiddlewareInterface
             ->withAttribute(self::MESSAGE_NAME_ATTRIBUTE, $name)
             ->withParsedBody($payload);
 
-        // Let's create a middleware pipeline of mapped middlewares
-        $middleware = $this->getMiddlewareForMessage($name);
-        $response   = $middleware->process($request, $delegate);
-
-        // Some middleware may return a 204 or any other 2xx answer, which are considered as success. However Elastic Beanstalk is picky
-        // and only accept 200 OK to delete message. So we normalize any status in the 2xx range
-        $statusCode = $response->getStatusCode();
-
-        if ($statusCode >= 200 && $statusCode <= 299) {
-            $response = $response->withStatus(200);
-        }
-
-        return $response->withHeader('X-Handled-By', 'ZfrEbWorker');
-    }
-
-    /**
-     * @param string $messageName
-     *
-     * @return MiddlewareInterface
-     * @throws RuntimeException
-     * @throws InvalidArgumentException
-     */
-    private function getMiddlewareForMessage(string $messageName): MiddlewareInterface
-    {
-        if (!array_key_exists($messageName, $this->messagesMapping)) {
-            throw new RuntimeException(sprintf(
-                'No middleware was mapped for message "%s". Did you fill the "zfr_eb_worker" configuration?',
-                $messageName
-            ));
-        }
-
-        $mappedMiddleware = $this->messagesMapping[$messageName];
-
-        if (!is_string($mappedMiddleware)) {
-            throw new InvalidArgumentException(sprintf(
-                'Mapped middleware must be a string, %s given.',
-                is_object($mappedMiddleware) ? get_class($mappedMiddleware) : gettype($mappedMiddleware)
-            ));
-        }
-
-        return $this->container->get($mappedMiddleware);
+        return $delegate->process($request, $delegate);
     }
 
     /**
